market_place/serializers.py
from rest_framework import serializers

from market_place import models


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Product
        fields = '__all__'
